+++ apps/designer/app/designer/shared/assets/update-state-assets.test.ts
@@ -6,7 +6,7 @@ import {
   DeletingClientAsset,
   ClientAsset,
 } from "./types";
-import { updateStateAssets } from "./update-asset-containers";
+import { updateStateAssets } from "./update-state-assets";
 
 const getAssetId = (clientAsset: ClientAsset): string =>
   clientAsset.status === "uploading"
